refactor(text-editor): move context into an internal directory
Moves `EditorContext` into an `__internal__` directory and updates imports where
neccessary. The context is not intended for public use and should improve tree-shaking in Carbon.

BREAKING CHANGE: moves `EditorContext` into an internal directory
