+++ src/components/text-editor/__internal__/editor.context.ts
@@ -1,0 +1,6 @@
+import React from "react";
+
+export default React.createContext<{
+  onLinkAdded?: (url: string) => void;
+  editMode?: boolean;
+}>({});
